+++ execution_engine_testing/test_support/src/lib.rs
@@ -149,18 +149,18 @@ pub static DEFAULT_GENESIS_CONFIG: Lazy<GenesisConfig> = Lazy::new(|| {
 pub static DEFAULT_CHAINSPEC_REGISTRY: Lazy<ChainspecRegistry> =
     Lazy::new(|| ChainspecRegistry::new_with_genesis(&[1, 2, 3], &[4, 5, 6]));
 // This static constant has been deprecated in favor of the Production counterpart
-// which uses costs tables and values which reflect values used by the Casper main-net.
-// #[deprecated]
-// /// Default [`RunGenesisRequest`].
-// pub static DEFAULT_RUN_GENESIS_REQUEST: Lazy<RunGenesisRequest> = Lazy::new(|| {
-//     RunGenesisRequest::new(
-//         *DEFAULT_GENESIS_CONFIG_HASH,
-//         *DEFAULT_PROTOCOL_VERSION,
-//         #[allow(deprecated)]
-//         DEFAULT_EXEC_CONFIG.clone(),
-//         DEFAULT_CHAINSPEC_REGISTRY.clone(),
-//     )
-// });
+// which uses costs tables and values which reflect values used by the Casper Mainnet.
+#[deprecated]
+/// Default [`RunGenesisRequest`].
+pub static DEFAULT_RUN_GENESIS_REQUEST: Lazy<RunGenesisRequest> = Lazy::new(|| {
+    RunGenesisRequest::new(
+        *DEFAULT_GENESIS_CONFIG_HASH,
+        *DEFAULT_PROTOCOL_VERSION,
+        #[allow(deprecated)]
+        DEFAULT_EXEC_CONFIG.clone(),
+        DEFAULT_CHAINSPEC_REGISTRY.clone(),
+    )
+});
 /// [`RunGenesisRequest`] instantiated using chainspec values.
 pub static PRODUCTION_RUN_GENESIS_REQUEST: Lazy<RunGenesisRequest> = Lazy::new(|| {
     ChainspecConfig::create_genesis_request_from_production_chainspec(